refactor: EnrollmentPolicyFactory 생성 - PolicyType의 역할 분리
